+++ index.html
@@ -230,6 +230,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
