+++ js/util.js
@@ -1,0 +1,14 @@
+export function getRandomPositiveInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+export function checkLength (checkedString, maxLength) {
+  return checkedString.length <= maxLength;
+}
+
+export const getRandomArrayElement = (elements) => elements[getRandomPositiveInteger(0, elements.length - 1)];
+
+export const generateNumbers = (start, end) => new Array(end - start + 1).fill().map(() => start++);
